Refactor user model retrieval
This commit refactors the way the user model is retrieved in the
organization-users.js file. Instead of using findAll, it now uses
query to filter users based on the company_id. This change was
necessary to improve the performance of the user retrieval process
and to ensure that only relevant users are returned (#1234).